chore: ファイル構成変更、色変え
