+++ frontend/src/features/game/components/Canvas.tsx
